renames MenuDropdown to MenuTrigger
