+++ components/src/NavBar/MenuTrigger.js
@@ -4,8 +4,9 @@ import PropTypes from "prop-types";
 import { Manager, Reference, Popper } from "react-popper";
 import theme from "ComponentsRoot/theme";
 import SubMenu from "./SubMenu";
+import SubMenuTrigger from "./SubMenuTrigger";
 import SubMenuLink from "./SubMenuLink";
-import { Text } from "ComponentsRoot";
+import Icon from "../Icon/Icon";
 
 const SubMenuItemsList = styled.ul({
   listStyle: "none",
@@ -13,21 +14,32 @@ const SubMenuItemsList = styled.ul({
   margin: "0",
 });
 
-const SubMenuDropdownButton = styled.button({
-  display: "block",
-  padding: `${theme.space.x1} ${theme.space.x2}`,
+const MenuTriggerButton = styled.button({
+  display: "inline-flex",
+  color: theme.colors.white,
+  border: "none",
+  backgroundColor: "transparent",
+  justifyContent: "center",
+  alignItems: "center",
+  textDecoration: "none",
+  verticalAlign: "middle",
+  lineHeight: theme.lineHeights.base,
+  transition: ".2s",
+  fontSize: `${theme.fontSizes.medium}`,
+  padding: `${theme.space.x1} ${theme.space.half} ${theme.space.x1} ${theme.space.x2}`,
+  borderRadius: theme.radii.medium,
+  [`${Icon}`]: {
+    color: theme.colors.lightGrey,
+  },
   "&:hover, &:focus": {
     outline: "none",
-    backgroundColor: theme.colors.lightGrey,
+    color: theme.colors.lightBlue,
+    backgroundColor: theme.colors.black,
+    cursor: "pointer",
   },
   "&:disabled": {
     opacity: ".5",
   },
-  border: "none",
-  backgroundColor: "transparent",
-  textDecoration: "none",
-  textAlign: "left",
-  cursor: "pointer",
 });
 
 const keyCode = Object.freeze({
@@ -45,10 +57,10 @@ const keyCode = Object.freeze({
   "DOWN": 40,
 });
 
-const isDropdown = subMenuItem => (subMenuItem.subMenuItems);
+const isDropdown = menuItem => (menuItem.subMenuItems);
 
 /* eslint-disable react/destructuring-assignment */
-class SubMenuDropdown extends React.Component {
+class MenuTrigger extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -85,7 +97,7 @@ class SubMenuDropdown extends React.Component {
     });
   }
 
-  SubMenuDropdownEventHandlers() {
+  menuDropdownEventHandlers() {
     return ({
       onClick: () => {
         this.showSubMenu();
@@ -115,27 +127,18 @@ class SubMenuDropdown extends React.Component {
       <Manager>
         <Reference>
           {({ ref }) => (
-            <li>
-              <SubMenuDropdownButton aria-haspopup="true" aria-expanded={ this.state.subMenuOpen } { ...this.props } { ...this.SubMenuDropdownEventHandlers() } ref={ ref }>
-                  <Text color={ "darkBlue" }>{ this.props.text }</Text>
-                    {this.props.subText && (
-                      <Text color={ "darkGrey" } fontSize={ theme.fontSizes.small } lineHeight={ theme.lineHeights.smallTextBase }>
-                    {this.props.subText}
-                  </Text>
-                  )}      
-              </SubMenuDropdownButton>
-            </li>
+            <MenuTriggerButton aria-haspopup="true" aria-expanded={ this.state.subMenuOpen } { ...this.props } { ...this.menuDropdownEventHandlers() } ref={ ref }>{ this.props.labelText }<Icon icon="downArrow" size="20px" p={ 2 } /></MenuTriggerButton>
           )}
         </Reference>
         {this.state.subMenuOpen && (
-        <Popper placement="right-start">
+        <Popper placement="bottom-start">
           {popperProps => (
-            <SubMenu renderArrow={false} popperProps={ popperProps } { ...this.subMenuEventHandlers() }>
+            <SubMenu popperProps={ popperProps } { ...this.subMenuEventHandlers() }>
               <SubMenuItemsList>
               {this.props.menuData.map(subMenuItem => {
                 if (isDropdown(subMenuItem)) {
                   return (
-                    <SubMenuDropdown key={ subMenuItem.text } text={ subMenuItem.text } subText={subMenuItem.subText} menuData={ subMenuItem.subMenuItems }/>      
+                    <SubMenuTrigger key={ subMenuItem.text } text={ subMenuItem.text } subText={subMenuItem.subText} menuData={ subMenuItem.subMenuItems }/>      
                   );
                 } else {
                   return (
@@ -155,17 +158,18 @@ class SubMenuDropdown extends React.Component {
 /* eslint-enable react/destructuring-assignment */
 
 
-SubMenuDropdown.propTypes = {
-  text: PropTypes.string.isRequired,
+MenuTrigger.propTypes = {
+  labelText: PropTypes.string.isRequired,
   menuData: PropTypes.arrayOf(PropTypes.shape({})),
   showDelay: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   hideDelay: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
 };
 
-SubMenuDropdown.defaultProps = {
+MenuTrigger.defaultProps = {
+  children: null,
   menuData: null,
   showDelay: "100",
   hideDelay: "350",
 };
 
-export default SubMenuDropdown;
+export default MenuTrigger;
